Create application_submission_retrieved table PR fixes
